exam_api/exam_api/users/urls.py
from django.conf.urls import url
from rest_framework.routers import DefaultRouter

from .views import UserViewSet

router = DefaultRouter()
router.register(r'users', UserViewSet)

urlpatterns = []

urlpatterns += router.urls
